Разрешение истории коммитов.
