+++ src/main/java/ru/practicum/shareit/ShareItApplication.java
@@ -1,0 +1,12 @@
+package ru.practicum.shareit;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ShareItApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ShareItApplication.class, args);
+    }
+}
