+++ src/main/java/ru/practicum/shareit/booking/Booking.java
@@ -1,0 +1,12 @@
+package ru.practicum.shareit.booking;
+
+import java.util.Date;
+
+public class Booking {
+    Long id;
+    Date start;
+    Date end;
+    Long item;
+    Long booker;
+    String status;
+}
