+++ src/main/java/ru/practicum/shareit/booking/Status.java
@@ -1,0 +1,8 @@
+package ru.practicum.shareit.booking;
+
+public enum Status {
+    WAITING,
+    APPROVED,
+    REJECTED,
+    CANCELED
+}
