+++ src/main/java/ru/practicum/shareit/exception/EmailAlreadyExistException.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.exception;
+
+public class EmailAlreadyExistException extends RuntimeException {
+    public EmailAlreadyExistException(final String message) {
+        super(message);
+    }
+}
