+++ src/main/java/ru/practicum/shareit/exception/EmptyEmailException.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.exception;
+
+public class EmptyEmailException extends RuntimeException {
+    public EmptyEmailException(final String message) {
+        super(message);
+    }
+}
