+++ src/main/java/ru/practicum/shareit/exception/ErrorResponse.java
@@ -1,0 +1,10 @@
+package ru.practicum.shareit.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class ErrorResponse {
+    private final String error;
+}
