+++ src/main/java/ru/practicum/shareit/exception/ItemNotFoundException.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.exception;
+
+public class ItemNotFoundException extends RuntimeException {
+    public ItemNotFoundException(final String message) {
+        super(message);
+    }
+}
