+++ src/main/java/ru/practicum/shareit/exception/UserNotFoundException.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.exception;
+
+public class UserNotFoundException extends RuntimeException {
+    public UserNotFoundException(final String message) {
+        super(message);
+    }
+}
