+++ src/main/java/ru/practicum/shareit/exception/WrongUserException.java
@@ -1,0 +1,8 @@
+package ru.practicum.shareit.exception;
+
+public class WrongUserException extends RuntimeException {
+    public WrongUserException(final String message) {
+        super(message);
+    }
+
+}
