+++ src/main/java/ru/practicum/shareit/item/Item.java
@@ -1,0 +1,16 @@
+package ru.practicum.shareit.item;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class Item {
+    Long id;
+    String name;
+    String description;
+    Boolean available;
+    long owner;
+    String request;
+
+}
