+++ src/main/java/ru/practicum/shareit/item/ItemDto.java
@@ -1,0 +1,16 @@
+package ru.practicum.shareit.item;
+
+import lombok.Value;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+
+@Value
+public class ItemDto {
+    @NotBlank(message = "Имя не указано")
+    String name;
+    @NotBlank(message = "Описание не может быть пустым")
+    String description;
+    @NotNull(message = "Доступность должна быть указана")
+    Boolean available;
+}
