+++ src/main/java/ru/practicum/shareit/item/ItemPatchDto.java
@@ -1,0 +1,11 @@
+package ru.practicum.shareit.item;
+
+import lombok.Value;
+
+@Value
+public class ItemPatchDto {
+    long id;
+    String name;
+    String description;
+    Boolean available;
+}
