+++ src/main/java/ru/practicum/shareit/item/ItemService.java
@@ -1,0 +1,56 @@
+package ru.practicum.shareit.item;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import ru.practicum.shareit.exception.ItemNotFoundException;
+import ru.practicum.shareit.exception.WrongUserException;
+import ru.practicum.shareit.item.repository.ItemRepository;
+import ru.practicum.shareit.user.User;
+import ru.practicum.shareit.user.UserService;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class ItemService {
+
+    private final ItemMapper itemMapper;
+    private final ItemRepository itemRepository;
+    private final UserService userService;
+
+    public Item addItem(long userId, ItemDto itemDto) {
+        User user = userService.getUserById(userId);
+        Item item = itemMapper.dtoToItem(itemDto);
+        item.setOwner(user.getId());
+        return itemRepository.addItem(item);
+    }
+
+    Item redactItem(long userId, long itemId, ItemPatchDto itemPatchDto) {
+        Item itemFromRepo = itemRepository.getItemById(itemId).orElseThrow(() -> new ItemNotFoundException("Вещь не найдена"));
+        if (itemFromRepo.getOwner() != userId) {
+            throw new WrongUserException("У вас нет доступа к данной вещи.");
+        }
+        Item item = itemMapper.patchDtoToItem(itemPatchDto);
+        return itemRepository.redactItem(itemId, item);
+    }
+
+    Item getItemById(long itemId) {
+        return itemRepository.getItemById(itemId)
+                .orElseThrow(() -> new ItemNotFoundException("Вещь с данным id не найдена"));
+    }
+
+    List<Item> searchItem(String text) {
+        if (text.isBlank()) {
+            return List.of();
+        }
+        return itemRepository.searchItem(text);
+    }
+
+    List<Item> getAllItemsOfUser(Long userId) {
+        return itemRepository.getAllItemsOfUser(userId);
+    }
+
+    public void deleteItem(long itemId) {
+        itemRepository.deleteItem(itemId);
+    }
+}
