+++ src/main/java/ru/practicum/shareit/request/ItemRequest.java
@@ -1,0 +1,10 @@
+package ru.practicum.shareit.request;
+
+import java.util.Date;
+
+public class ItemRequest {
+    Long id;
+    String description;
+    Long requester;
+    Date created;
+}
