+++ src/main/java/ru/practicum/shareit/user/User.java
@@ -1,0 +1,12 @@
+package ru.practicum.shareit.user;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class User {
+    Long id;
+    String name;
+    String email;
+}
