+++ src/main/java/ru/practicum/shareit/user/UserDto.java
@@ -1,0 +1,15 @@
+package ru.practicum.shareit.user;
+
+import lombok.Value;
+
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotBlank;
+
+@Value
+public class UserDto {
+    @NotBlank(message = "Имя пользователя не указано")
+    String name;
+    @Email(message = "Неправильный email")
+    @NotBlank(message = "Email не указан")
+    String email;
+}
