+++ src/main/java/ru/practicum/shareit/user/UserMapper.java
@@ -1,0 +1,9 @@
+package ru.practicum.shareit.user;
+
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface UserMapper {
+
+    User dtoToUser(UserDto user);
+}
