+++ src/main/java/ru/practicum/shareit/user/repository/UserRepository.java
@@ -1,0 +1,23 @@
+package ru.practicum.shareit.user.repository;
+
+import org.springframework.stereotype.Repository;
+import ru.practicum.shareit.user.User;
+
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+public interface UserRepository {
+
+    User createUser(User user);
+
+    User redactUser(long userId, User user);
+
+    Optional<User> getUserById(long userId);
+
+    Optional<User> getUserByEmail(String email);
+
+    List<User> getAllUsers();
+
+    void deleteUser(long userId);
+}
