--- src/main/java/ru/yandex/practicum/ShareItApplication.java
@@ -1,12 +1,0 @@
-package ru.yandex.practicum;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ShareItApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ShareItApplication.class, args);
-    }
-}
