--- src/main/java/ru/yandex/practicum/booking/Booking.java
@@ -1,12 +1,0 @@
-package ru.yandex.practicum.booking;
-
-import java.util.Date;
-
-public class Booking {
-    Long id;
-    Date start;
-    Date end;
-    Long item;
-    Long booker;
-    String status;
-}
