--- src/main/java/ru/yandex/practicum/booking/Status.java
@@ -1,8 +1,0 @@
-package ru.yandex.practicum.booking;
-
-public enum Status {
-    WAITING,
-    APPROVED,
-    REJECTED,
-    CANCELED
-}
