--- src/main/java/ru/yandex/practicum/exception/EmailAlreadyExistException.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.exception;
-
-public class EmailAlreadyExistException extends RuntimeException{
-    public EmailAlreadyExistException(final String message){
-        super(message);
-    }
-}
