--- src/main/java/ru/yandex/practicum/exception/EmptyEmailException.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.exception;
-
-public class EmptyEmailException extends RuntimeException{
-    public EmptyEmailException(final String message){
-        super(message);
-    }
-}
