--- src/main/java/ru/yandex/practicum/exception/ErrorResponse.java
@@ -1,10 +1,0 @@
-package ru.yandex.practicum.exception;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class ErrorResponse {
-    private final String error;
-}
