--- src/main/java/ru/yandex/practicum/exception/ItemNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.exception;
-
-public class ItemNotFoundException extends RuntimeException {
-    public ItemNotFoundException(final String message){
-        super(message);
-    }
-}
