--- src/main/java/ru/yandex/practicum/exception/UserNotFoundException.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.exception;
-
-public class UserNotFoundException extends RuntimeException{
-    public UserNotFoundException(final String message){
-        super(message);
-    }
-}
