--- src/main/java/ru/yandex/practicum/exception/WrongUserException.java
@@ -1,8 +1,0 @@
-package ru.yandex.practicum.exception;
-
-public class WrongUserException extends RuntimeException{
-    public WrongUserException(final String message){
-        super(message);
-    }
-
-}
