--- src/main/java/ru/yandex/practicum/item/Item.java
@@ -1,16 +1,0 @@
-package ru.yandex.practicum.item;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class Item {
-    Long id;
-    String name;
-    String description;
-    Boolean available;
-    long owner;
-    String request;
-
-}
