--- src/main/java/ru/yandex/practicum/item/ItemDto.java
@@ -1,16 +1,0 @@
-package ru.yandex.practicum.item;
-
-import lombok.Value;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-
-@Value
-public class ItemDto {
-    @NotBlank(message = "Имя не указано")
-    String name;
-    @NotBlank(message = "Описание не может быть пустым")
-    String description;
-    @NotNull(message = "Доступность должна быть указана")
-    Boolean available;
-}
