--- src/main/java/ru/yandex/practicum/item/ItemPatchDto.java
@@ -1,11 +1,0 @@
-package ru.yandex.practicum.item;
-
-import lombok.Value;
-
-@Value
-public class ItemPatchDto {
-    long id;
-    String name;
-    String description;
-    Boolean available;
-}
