--- src/main/java/ru/yandex/practicum/request/ItemRequest.java
@@ -1,10 +1,0 @@
-package ru.yandex.practicum.request;
-
-import java.util.Date;
-
-public class ItemRequest {
-    Long id;
-    String description;
-    Long requester;
-    Date created;
-}
