--- src/main/java/ru/yandex/practicum/user/User.java
@@ -1,12 +1,0 @@
-package ru.yandex.practicum.user;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class User {
-    Long id;
-    String name;
-    String email;
-}
