--- src/main/java/ru/yandex/practicum/user/UserController.java
@@ -1,44 +1,0 @@
-package ru.yandex.practicum.user;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
-
-import javax.validation.Valid;
-import java.util.List;
-
-@Slf4j
-@RestController()
-@RequiredArgsConstructor
-@RequestMapping("/users")
-public class UserController {
-    private final UserService userService;
-
-    @PostMapping
-    @ResponseStatus(HttpStatus.CREATED)
-    public User createUser(@Valid @RequestBody UserDto userDto){
-        return userService.createUser(userDto);
-    }
-
-    @GetMapping("/{userId}")
-    public User getUserById(@PathVariable long userId){
-        return userService.getUserById(userId);
-    }
-
-    @GetMapping
-    public List<User> getAllUsers(){
-        return userService.getAllUsers();
-    }
-
-    @PatchMapping("/{userId}")
-    public User redactUserInfo (@PathVariable long userId,
-                                @Valid @RequestBody UserPatchDto userPatchDto){
-        return userService.redactUser(userId, userPatchDto);
-    }
-
-    @DeleteMapping("/{userId}")
-    public void deleteUserById (@PathVariable long userId){
-        userService.deleteUser(userId);
-    }
-}
