--- src/main/java/ru/yandex/practicum/user/UserDto.java
@@ -1,15 +1,0 @@
-package ru.yandex.practicum.user;
-
-import lombok.Value;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-
-@Value
-public class UserDto {
-    @NotBlank(message = "Имя пользователя не указано")
-    String name;
-    @Email(message = "Неправильный email")
-    @NotBlank(message = "Email не указан")
-    String email;
-}
