--- src/main/java/ru/yandex/practicum/user/UserMapper.java
@@ -1,9 +1,0 @@
-package ru.yandex.practicum.user;
-
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel = "spring")
-public interface UserMapper {
-
-    User dtoToUser(UserDto user);
-}
