--- src/main/java/ru/yandex/practicum/user/UserPatchDto.java
@@ -1,12 +1,0 @@
-package ru.yandex.practicum.user;
-
-import lombok.Value;
-
-import javax.validation.constraints.Email;
-
-@Value
-public class UserPatchDto {
-    String name;
-    @Email(message = "Неправильный email")
-    String email;
-}
